requirements.txt
Flask==2.1.1
twilio==7.2.0
groq==1.0.0
python-dotenv==0.19.2